# src/main/resources/static/docs/trycatch_openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "trycatch API",
    "description" : "trycatch API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://trycatch"
  } ],
  "paths" : {
...
    "/<unknown>/user/detail" : {
      "patch" : {
        "summary" : "PATCH /v1/user/detail",
        "operationId" : "patchUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserModifyDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/detail/{targetId}" : {
      "get" : {
        "summary" : "GET /v1/user/detail/{targetId}",
        "operationId" : "findUser",
        "parameters" : [ {
          "name" : "targetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "DELETE /v1/user/detail/{targetId}",
        "operationId" : "removeUser",
        "parameters" : [ {
          "name" : "targetId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WithdrawalRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/follow/{uid}" : {
      "put" : {
        "summary" : "PUT /v1/user/follow/{uid}",
        "operationId" : "unfollowUser",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "POST /v1/user/follow/{uid}",
        "operationId" : "followUser",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/id/{userName}" : {
      "get" : {
        "summary" : "GET /v1/user/id/{userName}",
        "operationId" : "findUserId",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/image/{userId}" : {
      "get" : {
        "summary" : "GET /v1/user/image/{userId}",
        "operationId" : "findUserImage",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/<unknown>/user/verification/{uid}" : {
      "get" : {
        "summary" : "GET /v1/user/verification/{uid}",
        "operationId" : "isConfirmed",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "POST /v1/user/verification/{uid}",
        "operationId" : "verifyCompany",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{uid}/answer/list" : {
      "get" : {
        "summary" : "GET /v1/user/{uid}/answer/list",
        "operationId" : "findUserAnswers",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{uid}/list" : {
      "get" : {
        "summary" : "GET /v1/user/{uid}/list",
        "operationId" : "findFollows",
        "parameters" : [ {
          "name" : "uid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userId}/badge/list" : {
      "get" : {
        "summary" : "GET /v1/user/{userId}/badge/list",
        "operationId" : "findBadge",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userId}/history" : {
      "get" : {
        "summary" : "GET /v1/user/{userId}/history",
        "operationId" : "findHistory",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userId}/question/list" : {
      "get" : {
        "summary" : "GET /v1/user/{userId}/question/list",
        "operationId" : "findUserQuestions",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userId}/recent/list" : {
      "get" : {
        "summary" : "GET /v1/user/{userId}/recent/list",
        "operationId" : "findRecentFeed",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userId}/subscription/list" : {
      "get" : {
        "summary" : "GET /v1/user/{userId}/subscription/list",
        "operationId" : "findSubscriptionList",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userName}/github/fetch" : {
      "get" : {
        "summary" : "GET /v1/user/{userName}/github/fetch",
        "operationId" : "fetchGitHub",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userName}/tag" : {
      "post" : {
        "summary" : "POST /v1/user/{userName}/tag",
        "operationId" : "addTag",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
    "/<unknown>/user/{userName}/tag/{tagId}" : {
      "delete" : {
        "summary" : "DELETE /v1/user/{userName}/tag/{tagId}",
        "operationId" : "removeTag",
        "parameters" : [ {
          "name" : "userName",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tagId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "HttpHeaders" : {
        "type" : "object",
        "properties" : { }
      },
      "ResponseEntity" : {
        "type" : "object",
        "properties" : {
          "headers" : {
            "$ref" : "#/components/schemas/HttpHeaders"
          }
        }
      },
...
      "UserModifyDto" : {
        "type" : "object",
        "properties" : {
          "introduction" : {
            "type" : "string"
          },
          "profileImage" : {
            "type" : "string"
          }
        }
      },
      "WithdrawalRequestDto" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "content" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}